Conflicts:
	src/main/java/cz/muni/fi/PA165/daoDtoConversion/BuildingKitConversion.java
	src/main/java/cz/muni/fi/PA165/service/BrickServiceImpl.java
	src/main/java/cz/muni/fi/PA165/service/BuildingKitService.java
	src/main/java/cz/muni/fi/PA165/service/CategoryServiceImpl.java
	src/test/java/cz/muni/fi/PA165/service/BuildingKitServiceTest.java
	src/test/java/cz/muni/fi/PA165/service/CategoryServiceTest.java
	src/test/java/cz/muni/fi/PA165/service/ThemeSetServiceTest.java

+++ src/main/java/cz/muni/fi/PA165/service/BrickServiceImpl.java
@@ -10,10 +10,18 @@ import cz.muni.fi.PA165.daoDtoConversion.BrickConversion;
 import cz.muni.fi.PA165.dto.BrickDto;
 import cz.muni.fi.PA165.entity.Brick;
 import cz.muni.fi.PA165.entity.Color;
+
 import java.util.ArrayList;
 import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
+import org.springframework.dao.DataAccessException;
 
 /**
  *
@@ -23,7 +31,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class BrickServiceImpl implements BrickService {
 
-    //@Autowired
+    @Autowired
     private BrickDao brickDao;
 
     public void setBrickDao(BrickDao brickDao) {
@@ -31,16 +39,25 @@ public class BrickServiceImpl implements BrickService {
     }
 
     public void create(BrickDto brickDto) {
+        if(brickDto == null) {
+            throw new DataAccessExceptionService("created brick cannot be null");
+        }
         Brick brickEntity = BrickConversion.convertToEntity(brickDto);
         brickDao.create(brickEntity);
     }
 
     public void update(BrickDto brickDto) {
+        if(brickDto == null) {
+            throw new DataAccessExceptionService("updated brick cannot be null");
+        }
         Brick brickEntity = BrickConversion.convertToEntity(brickDto);
         brickDao.update(brickEntity);
     }
 
     public void delete(Long id) {
+        if(id == null) {
+            throw new DataAccessExceptionService("cannot delete brick with null id");
+        }
         brickDao.delete(id);
     }
 
@@ -54,11 +71,17 @@ public class BrickServiceImpl implements BrickService {
     }
 
     public BrickDto findById(Long id) {
+        if(id == null) {
+            throw new DataAccessExceptionService("cannot find brick with null id");
+        }
         Brick entity = brickDao.findById(id);
         return BrickConversion.convertToDto(entity);
     }
 
     public List<BrickDto> findByColor(Color color) {
+        if(color == null) {
+            throw new DataAccessExceptionService("cannot find brick with null color");
+        }
         List<Brick> brickEntitites = brickDao.findByColor(color);
         List<BrickDto> brickDtoList = new ArrayList<BrickDto>();
         for (Brick brick : brickEntitites) {
@@ -68,6 +91,9 @@ public class BrickServiceImpl implements BrickService {
     }
 
     public List<BrickDto> findByName(String name) {
+        if(name == null) {
+            throw new DataAccessExceptionService("cannot brick with null name");
+        }
         List<Brick> brickEntitites = brickDao.findByName(name);
         List<BrickDto> brickDtoList = new ArrayList<BrickDto>();
         for (Brick brick : brickEntitites) {
